Required SoapClient for all integrations

+++ includes/integrations/class-wc-correios-integration.php
@@ -20,8 +20,8 @@ class WC_Correios_Integration extends WC_Integration {
 	 * Initialize integration actions.
 	 */
 	public function __construct() {
-		$this->id                 = 'correios-integration';
-		$this->method_title       = __( 'Correios', 'woocommerce-correios' );
+		$this->id           = 'correios-integration';
+		$this->method_title = __( 'Correios', 'woocommerce-correios' );
 
 		// Load the form fields.
 		$this->init_form_fields();
@@ -87,65 +87,60 @@ class WC_Correios_Integration extends WC_Integration {
 				'default'     => 'no',
 				'description' => sprintf( __( 'Log %s events, such as WebServices requests.', 'woocommerce-correios' ), __( 'Tracking History Table', 'woocommerce-correios' ) ) . $this->get_tracking_log_link(),
 			),
+			'autofill_addresses' => array(
+				'title'       => __( 'Autofill Addresses', 'woocommerce-correios' ),
+				'type'        => 'title',
+				'description' => __( 'Displays a table with informations about the shipping in My Account > View Order page.', 'woocommerce-correios' ),
+			),
+			'autofill_enable' => array(
+				'title'   => __( 'Enable/Disable', 'woocommerce-correios' ),
+				'type'    => 'checkbox',
+				'label'   => __( 'Enable Autofill Addresses', 'woocommerce-correios' ),
+				'default' => 'no',
+			),
+			'autofill_validity' => array(
+				'title'       => __( 'Postcodes Validity', 'woocommerce-correios' ),
+				'type'        => 'select',
+				'default'     => 'forever',
+				'class'       => 'wc-enhanced-select',
+				'description' => __( 'Defines how long a postcode will stay saved in the database before a new query.', 'woocommerce-correios' ),
+				'options'     => array(
+					'1'     => __( '1 month', 'woocommerce-correios' ),
+					'2'     => sprintf( __( '%d month', 'woocommerce-correios' ), 2 ),
+					'3'     => sprintf( __( '%d month', 'woocommerce-correios' ), 3 ),
+					'4'     => sprintf( __( '%d month', 'woocommerce-correios' ), 4 ),
+					'5'     => sprintf( __( '%d month', 'woocommerce-correios' ), 5 ),
+					'6'     => sprintf( __( '%d month', 'woocommerce-correios' ), 6 ),
+					'7'     => sprintf( __( '%d month', 'woocommerce-correios' ), 7 ),
+					'8'     => sprintf( __( '%d month', 'woocommerce-correios' ), 8 ),
+					'9'     => sprintf( __( '%d month', 'woocommerce-correios' ), 9 ),
+					'10'    => sprintf( __( '%d month', 'woocommerce-correios' ), 10 ),
+					'11'    => sprintf( __( '%d month', 'woocommerce-correios' ), 11 ),
+					'12'    => sprintf( __( '%d month', 'woocommerce-correios' ), 12 ),
+					'forever' => __( 'Forever', 'woocommerce-correios' ),
+				),
+			),
+			'autofill_force' => array(
+				'title'       => __( 'Force Autofill', 'woocommerce-correios' ),
+				'type'        => 'checkbox',
+				'label'       => __( 'Enable Force Autofill', 'woocommerce-correios' ),
+				'description' => __( 'When enabled will autofill all addresses after the user finish to fill the postcode, even if the addresses are already filled.', 'woocommerce-correios' ),
+				'default'     => 'no',
+			),
+			'autofill_empty_database' => array(
+				'title'       => __( 'Empty Database', 'woocommerce-correios' ),
+				'type'        => 'button',
+				'label'       => __( 'Empty Database', 'woocommerce-correios' ),
+				'description' => __( 'Delete all the saved postcodes in the database, use this option if you have issues with outdated postcodes.', 'woocommerce-correios' ),
+			),
+			'autofill_debug' => array(
+				'title'       => __( 'Debug Log', 'woocommerce-correios' ),
+				'type'        => 'checkbox',
+				'label'       => __( 'Enable logging for Autofill Addresses', 'woocommerce-correios' ),
+				'default'     => 'no',
+				'description' => sprintf( __( 'Log %s events, such as WebServices requests.', 'woocommerce-correios' ), __( 'Autofill Addresses', 'woocommerce-correios' ) ) . $this->get_tracking_log_link(),
+			),
 		);
-
-		if ( $this->check_if_support_autofill() ) {
-			$this->form_fields += array(
-				'autofill_addresses' => array(
-					'title'       => __( 'Autofill Addresses', 'woocommerce-correios' ),
-					'type'        => 'title',
-					'description' => __( 'Displays a table with informations about the shipping in My Account > View Order page.', 'woocommerce-correios' ),
-				),
-				'autofill_enable' => array(
-					'title'   => __( 'Enable/Disable', 'woocommerce-correios' ),
-					'type'    => 'checkbox',
-					'label'   => __( 'Enable Autofill Addresses', 'woocommerce-correios' ),
-					'default' => 'no',
-				),
-				'autofill_validity' => array(
-					'title'       => __( 'Postcodes Validity', 'woocommerce-correios' ),
-					'type'        => 'select',
-					'default'     => 'forever',
-					'class'       => 'wc-enhanced-select',
-					'description' => __( 'Defines how long a postcode will stay saved in the database before a new query.', 'woocommerce-correios' ),
-					'options'     => array(
-						'1'     => __( '1 month', 'woocommerce-correios' ),
-						'2'     => sprintf( __( '%d month', 'woocommerce-correios' ), 2 ),
-						'3'     => sprintf( __( '%d month', 'woocommerce-correios' ), 3 ),
-						'4'     => sprintf( __( '%d month', 'woocommerce-correios' ), 4 ),
-						'5'     => sprintf( __( '%d month', 'woocommerce-correios' ), 5 ),
-						'6'     => sprintf( __( '%d month', 'woocommerce-correios' ), 6 ),
-						'7'     => sprintf( __( '%d month', 'woocommerce-correios' ), 7 ),
-						'8'     => sprintf( __( '%d month', 'woocommerce-correios' ), 8 ),
-						'9'     => sprintf( __( '%d month', 'woocommerce-correios' ), 9 ),
-						'10'    => sprintf( __( '%d month', 'woocommerce-correios' ), 10 ),
-						'11'    => sprintf( __( '%d month', 'woocommerce-correios' ), 11 ),
-						'12'    => sprintf( __( '%d month', 'woocommerce-correios' ), 12 ),
-						'forever' => __( 'Forever', 'woocommerce-correios' ),
-					),
-				),
-				'autofill_force' => array(
-					'title'       => __( 'Force Autofill', 'woocommerce-correios' ),
-					'type'        => 'checkbox',
-					'label'       => __( 'Enable Force Autofill', 'woocommerce-correios' ),
-					'description' => __( 'When enabled will autofill all addresses after the user finish to fill the postcode, even if the addresses are already filled.', 'woocommerce-correios' ),
-					'default'     => 'no',
-				),
-				'autofill_empty_database' => array(
-					'title'       => __( 'Empty Database', 'woocommerce-correios' ),
-					'type'        => 'button',
-					'label'       => __( 'Empty Database', 'woocommerce-correios' ),
-					'description' => __( 'Delete all the saved postcodes in the database, use this option if you have issues with outdated postcodes.', 'woocommerce-correios' ),
-				),
-				'autofill_debug' => array(
-					'title'       => __( 'Debug Log', 'woocommerce-correios' ),
-					'type'        => 'checkbox',
-					'label'       => __( 'Enable logging for Autofill Addresses', 'woocommerce-correios' ),
-					'default'     => 'no',
-					'description' => sprintf( __( 'Log %s events, such as WebServices requests.', 'woocommerce-correios' ), __( 'Autofill Addresses', 'woocommerce-correios' ) ) . $this->get_tracking_log_link(),
-				),
-			);
-		}
 	}
 
 	/**
@@ -157,8 +152,13 @@ class WC_Correios_Integration extends WC_Integration {
 
 		include WC_Correios::get_plugin_path() . 'includes/admin/views/html-admin-help-message.php';
 
-		echo '<div><input type="hidden" name="section" value="' . esc_attr( $this->id ) . '" /></div>';
-		echo '<table class="form-table">' . $this->generate_settings_html( $this->get_form_fields(), false ) . '</table>';
+		if ( class_exists( 'SoapClient' ) ) {
+			echo '<div><input type="hidden" name="section" value="' . esc_attr( $this->id ) . '" /></div>';
+			echo '<table class="form-table">' . $this->generate_settings_html( $this->get_form_fields(), false ) . '</table>';
+		} else {
+			$GLOBALS['hide_save_button'] = true; // Hide save button.
+			echo '<div class="notice notice-error inline"><p>' . sprintf( __( 'It\'s required have installed the %s on your server in order to integrate with the services of the Correios!', 'woocommerce-correios' ), '<a href="https://secure.php.net/manual/book.soap.php" target="_blank">' . __( 'SOAP module', 'woocommerce-correios' ) . '</a>' ) . '</p></div>';
+		}
 
 		$suffix = defined( 'SCRIPT_DEBUG' ) && SCRIPT_DEBUG ? '' : '.min';
 		wp_enqueue_script( $this->id . '-admin', plugins_url( 'assets/js/admin/integration' . $suffix . '.js', WC_Correios::get_main_file() ), array( 'jquery', 'jquery-blockui' ), WC_Correios::VERSION, true );
@@ -216,7 +216,7 @@ class WC_Correios_Integration extends WC_Integration {
 	 * @return bool
 	 */
 	public function setup_tracking_history() {
-		return 'yes' === $this->tracking_enable;
+		return 'yes' === $this->tracking_enable && class_exists( 'SoapClient' );
 	}
 
 	/**
@@ -229,21 +229,12 @@ class WC_Correios_Integration extends WC_Integration {
 	}
 
 	/**
-	 * Check if support autofill.
-	 *
-	 * @return bool
-	 */
-	protected function check_if_support_autofill() {
-		return defined( 'WC_VERSION' ) && version_compare( WC_VERSION, '2.5.0', '>=' ) && class_exists( 'SoapClient' );
-	}
-
-	/**
 	 * Enable autofill addresses.
 	 *
 	 * @return bool
 	 */
 	public function setup_autofill_addresses() {
-		return 'yes' === $this->autofill_enable && $this->check_if_support_autofill();
+		return 'yes' === $this->autofill_enable && class_exists( 'SoapClient' );
 	}
 
 	/**
